cleanup; added XCIntegrator.getTimingSummary

+++ src/dft/DensityGridGenerator.cpp
@@ -140,7 +140,7 @@ generateDensityForGGA(double*             rho,
         auto grid_batch_offset = mathfunc::batch_offset(npoints, thread_id, nthreads);
 
 #pragma omp simd
-        for (int32_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
+        for (int64_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
         {
             // rho_a
             rho[2 * g + 0] = 0.0;
@@ -151,12 +151,12 @@ generateDensityForGGA(double*             rho,
             rhograd[6 * g + 2] = 0.0;
         }
 
-        for (int32_t nu = 0; nu < naos; nu++)
+        for (int64_t nu = 0; nu < naos; nu++)
         {
             auto nu_offset = nu * npoints;
 
 #pragma omp simd
-            for (int32_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
+            for (int64_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
             {
                 // rho_a
                 rho[2 * g + 0] += F_val[nu_offset + g] * chi_val[nu_offset + g];
@@ -169,7 +169,7 @@ generateDensityForGGA(double*             rho,
         }
 
 #pragma omp simd
-        for (int32_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
+        for (int64_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
         {
             // rho_b
             rho[2 * g + 1] = rho[2 * g + 0];
@@ -183,7 +183,7 @@ generateDensityForGGA(double*             rho,
         if (sigma != nullptr)
         {
 #pragma omp simd
-            for (int32_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
+            for (int64_t g = grid_batch_offset; g < grid_batch_offset + grid_batch_size; g++)
             {
                 // simga_aa, sigma_ab, sigma_bb
                 // clang-format off
